[REMANIEMENT][TESTS POSTGRES] Rassemble les tests sur les entrepôts Postgres au sein d’un même fichier de tests pour éviter les effets de bords entre les tests
